nonebot_plugin_eop_ai/pyproject.toml
[tool.poetry]
name = "nonebot_plugin_eop_ai"
version = "0.1.0"
description = "Nonebot2 一款调用eop api的AI聊天插件"
authors = ["nikissXI <1299577815@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_eop_ai"}]
homepage = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_eop_ai"
repository = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_eop_ai"
documentation = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_eop_ai#README.md"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = ">=3.8"
nonebot2 = ">=2.0.0rc1"
nonebot-adapter-onebot = ">=2.0.0"
httpx = ">=0.24.0"
nonebot-plugin-htmlrender = ">=0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
